+++ changelog.md
@@ -1,4 +1,7 @@
 ### 0.0.1
 
 * Code migrated into this repo
+* Created some tests
+* Get pull out the class methods from the class
+